chore: split combat modules
