+++ docs/deep-dives/extractors-layer.md
@@ -20,7 +20,7 @@ Prerequisites:
 3. [Parser Layer (`parser.py`)](parser-layer.md)
 4. [SendTable Layer (`schema/sendtable.py`)](sendtable-layer.md)
 5. [State Reconstruction Layer (`state/string_table.py` + `state/entities.py`)](state-layer.md)
-6. [Event Normalization Layer (`state/game_events.py` + `combatlog.py`)](event-layer.md)
+6. [Event Normalization Layer (`state/game_events.py` + `combat/log.py`)](event-layer.md)
 
 ## Why this is the next layer
