+++ docs/reference/combat_aggregator.md
@@ -6,10 +6,10 @@ Aggregates combat log entries into per-player statistics.
 
 ## Generated API
 
-## Module `gem.combat_aggregator`
+## Module `gem.combat.aggregator`
 
 Per-player combat log aggregation for gem replay parsing.
 
-Source: [src/gem/combat_aggregator.py](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/combat_aggregator.py#L1)
+Source: [src/gem/combat/aggregator.py](https://github.com/whanyu1212/gem-dota/blob/main/src/gem/combat/aggregator.py#L1)
 
 No public top-level classes or functions found.
