+++ examples/extraction_demo.py
@@ -21,7 +21,7 @@ from pathlib import Path
 # Allow running from repo root without installing
 sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 
-from gem.combatlog import CombatLogEntry
+from gem.combat.log import CombatLogEntry
 from gem.constants import (
     ABILITIES,
     HEROES,
