+++ src/gem/combat/__init__.py
@@ -1,0 +1,14 @@
+"""Combat log ingestion and aggregation primitives."""
+
+from gem.combat.aggregator import _CombatAggregator, _dedup_purchase_log, _ParsedPlayerAgg
+from gem.combat.log import COMBAT_LOG_TYPES, CombatLogEntry, CombatLogHandler, CombatLogProcessor
+
+__all__ = [
+    "COMBAT_LOG_TYPES",
+    "CombatLogEntry",
+    "CombatLogHandler",
+    "CombatLogProcessor",
+    "_CombatAggregator",
+    "_ParsedPlayerAgg",
+    "_dedup_purchase_log",
+]
