+++ src/gem/extractors/wards.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Literal
 
-from gem.combatlog import CombatLogEntry
+from gem.combat.log import CombatLogEntry
 from gem.state.entities import Entity, EntityOp
 
 if TYPE_CHECKING:
